Account management added, Edit/Delete logic for reviews added. Automatically determine what business is linked to the review when writing + other minor changes

+++ BusinessReviewApp/Controllers/ReviewController.cs
@@ -13,14 +13,14 @@ namespace BusinessReviewApp.Controllers
     {
         private UsersContext db = new UsersContext();
 
-        //
+        // Currently disabled as there is no need for it
         // GET: /Review/
 
-        public ActionResult Index()
+        /*public ActionResult Index()
         {
             var reviews = db.Reviews.Include(r => r.Business).Include(r => r.UserProfile);
             return View(reviews.ToList());
-        }
+        }*/
 
         //
         // GET: /Review/Details/5
@@ -35,16 +35,12 @@ namespace BusinessReviewApp.Controllers
             return View(review);
         }
 
-        //
-        // GET: /Review/Create
-         [Authorize]
-        public ActionResult Create()
+        [Authorize]
+        public ActionResult Create(int id)
         {
-            ViewBag.BusinessID = new SelectList(db.Businesses, "BusinessID", "Name");
-            ViewBag.UserId = new SelectList(db.UserProfiles, "UserId", "UserName");
-            return View();
+            ViewBag.businessID = id;
+            return View(new Review());
         }
-
         //
         // POST: /Review/Create
         [Authorize]
@@ -72,7 +68,7 @@ namespace BusinessReviewApp.Controllers
                     }
                 }
 
-                //Assume business review for that user already exists
+                //Assume business review for that user doesnt already exists
                 bool reviewAlreadyExists = false;
 
                 //Check to see if there is already a review for the business for the user.
@@ -81,10 +77,10 @@ namespace BusinessReviewApp.Controllers
                 //Foreach Review check if there is reviews for the business
                 foreach (var item in db.Reviews)
                 {
-                    if(item.BusinessID == review.BusinessID)
+                    if (item.BusinessID == review.BusinessID)
                     {
                         reviews.Add(item);
-                    }    
+                    }
                 }
 
                 //Foreach review, check if the user has already added one.
@@ -93,13 +89,13 @@ namespace BusinessReviewApp.Controllers
                     if (item.UserId == review.UserId)
                     {
                         reviewAlreadyExists = true;
-                    }              
+                    }
                 }
 
                 //Check if a review for this business from this user does not already exist
-                if(reviewAlreadyExists == true)
+                if (reviewAlreadyExists == true)
                 {
-                    ModelState.AddModelError("", "You have already review this business and may not add another review.");
+                    ModelState.AddModelError("", "You have already reviewed this business and may not add another review.");
                 }
                 else
                 {
@@ -110,53 +106,24 @@ namespace BusinessReviewApp.Controllers
                     //Update Rating for business
                     calculateRating(review);
 
-
-
-
-                    BusinessReviewViewModel businessReviewVM = new BusinessReviewViewModel();
-                    businessReviewVM.businesses = db.Businesses.Find(review.BusinessID);
-                    if (businessReviewVM.businesses == null)
-                    {
-                        return HttpNotFound();
-                    }
-
-                    businessReviewVM.reviews = new List<Review>();
-                    foreach (var item in db.Reviews.ToList())
-                    {
-                        if (item.BusinessID == db.Businesses.Find(review.BusinessID).BusinessID)
-                        {
-                            businessReviewVM.reviews.Add(item);
-                        }
-                    }
-                    return View(businessReviewVM);
-
-
-
-
-
-
-                    //return RedirectToAction("Index");
+                    return RedirectToAction("Index", "Business");
                 }
-               
             }
-
-            ViewBag.BusinessID = new SelectList(db.Businesses, "BusinessID", "Name", review.BusinessID);
-            ViewBag.UserId = new SelectList(db.UserProfiles, "UserId", "UserName", review.UserId);
+            ViewBag.businessID = review.BusinessID;
             return View(review);
         }
 
         //
         // GET: /Review/Edit/5
         [Authorize]
-        public ActionResult Edit(int id = 0)
+        public ActionResult Edit(int id)
         {
             Review review = db.Reviews.Find(id);
             if (review == null)
             {
                 return HttpNotFound();
             }
-            ViewBag.BusinessID = new SelectList(db.Businesses, "BusinessID", "Name", review.BusinessID);
-            ViewBag.UserId = new SelectList(db.UserProfiles, "UserId", "UserName", review.UserId);
+            ViewBag.BusinessID = id;
             return View(review);
         }
 
@@ -167,19 +134,37 @@ namespace BusinessReviewApp.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit(Review review)
         {
+            bool isYourReview = false;
             if (ModelState.IsValid)
             {
-                db.Entry(review).State = EntityState.Modified;
-                db.SaveChanges();
+                //Gets the currently logged in user
+                foreach (var item in db.UserProfiles)
+                {
+                    if (item.UserName == User.Identity.Name)
+                    {
+                        //checks to see if they wrote the review
+                        if (item.UserId == review.UserId)
+                        {
+                            isYourReview = true;
+                        }
 
-                //Update Rating for business
-                calculateRating(review);
-                
-                return RedirectToAction("Index");
+                    }
+                }
+                if (isYourReview == true)
+                {
+                    db.Entry(review).State = EntityState.Modified;
+                    db.SaveChanges();
+
+                    //Update Rating for business
+                    calculateRating(review);
+
+                    return RedirectToAction("Index", "Business");
+                }
+
             }
-            ViewBag.BusinessID = new SelectList(db.Businesses, "BusinessID", "Name", review.BusinessID);
-            ViewBag.UserId = new SelectList(db.UserProfiles, "UserId", "UserName", review.UserId);
-            return View(review);
+            ModelState.AddModelError("", "You can only edit reviews that you wrote.");
+            ViewBag.businessID = review.BusinessID;
+            return View();
         }
 
         //
@@ -187,6 +172,7 @@ namespace BusinessReviewApp.Controllers
         [Authorize]
         public ActionResult Delete(int id = 0)
         {
+
             Review review = db.Reviews.Find(id);
             if (review == null)
             {
@@ -202,14 +188,32 @@ namespace BusinessReviewApp.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            bool isYourReview = false;
             Review review = db.Reviews.Find(id);
-            db.Reviews.Remove(review);
-            db.SaveChanges();
 
-            //Update Rating for business
-            calculateRating(review);
+            //Gets the currently logged in user
+            foreach (var item in db.UserProfiles)
+            {
+                if (item.UserName == User.Identity.Name)
+                {
+                    //checks to see if they wrote the review
+                    if (item.UserId == review.UserId)
+                    {
+                        isYourReview = true;
+                    }
+                 }
+            }
+            if (isYourReview == true)
+            {
+                db.Reviews.Remove(review);
+                db.SaveChanges();
 
-            return RedirectToAction("Index");
+                //Update Rating for business
+                calculateRating(review);
+                return RedirectToAction("Index", "Business");
+            }
+            ModelState.AddModelError("", "You may only delete reviews you wrote.");
+            return View(review);
         }
 
         protected override void Dispose(bool disposing)
@@ -238,7 +242,7 @@ namespace BusinessReviewApp.Controllers
 
             Business b1 = new Business();
             b1 = db.Businesses.Find(review.BusinessID);
-            
+
             //Check if there are no reviews to ensure there is no divide by 0 exception
             if (totalNumberOfReviews == 0)
             {
